requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# MQTT
paho-mqtt==1.6.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Async
aiofiles==23.2.1

# Analytics & Charts
matplotlib==3.8.0
pandas==2.1.3
openpyxl==3.1.2

# Scheduling
croniter==2.0.1

# PC Monitor (Windows)
psutil==5.9.0
pythonnet==3.0.5
wmi==1.5.1
